requirements.txt
pandas==1.2.0
fire==0.4.0
requests==2.23.0
tensorflow==2.2.0
Keras==2.4.3
starlette==0.14.2
pypandoc==1.5
uvicorn==0.13.4
numpy==1.19.5
pillow==8.1.2
scikit_learn==0.24.1
scrapy==2.1.0
gevent==20.6.2
h5py==2.10.0
aiofiles==0.5.0
aiohttp==3.6.2
asyncio==3.4.3
python-multipart==0.0.5